sort button added
sort button added

+++ src/views/apiCheck/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import Table from "../../components/Table";
 import InputForm from "../../components/inputForm";
+import Loader from "../../components/loader";
 
 export default class ApiTest extends React.Component {
   constructor(props) {
@@ -29,9 +30,17 @@ export default class ApiTest extends React.Component {
 
   handleChange = e => {
     if (e.target.id === "name" || e.target.id === "newName") {
+      if (e.target.value.match(/^[A-Za-z]+$/)) {
+      }
     } else if (e.target.id === "username") {
     } else if (e.target.id === "email") {
+      if (
+        e.target.value.match(/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/)
+      ) {
+      }
     } else if (e.target.id === "phone" || e.target.id === "newPhone") {
+      if (e.target.value.match(/^[-+]?[0-9]+$/)) {
+      }
     }
     this.setState({
       [e.target.id]: e.target.value
@@ -56,19 +65,31 @@ export default class ApiTest extends React.Component {
     })
       .then(res => res.json())
       .then(data => {
-        alert(data);
         console.log(data);
       })
       .catch(console.log);
     /*******************************************************/
     let { data } = this.state;
-    data.push({
-      id: data[data.length - 1].id + 1,
-      name: this.state.name,
-      username: this.state.username,
-      email: this.state.email,
-      phone: this.state.phone
-    });
+    data.splice(
+      data.findIndex(
+        ele =>
+          ele.id ===
+          data.reduce((prev, current) =>
+            prev.id > current.id ? prev : current
+          ).id
+      ),
+      0,
+      {
+        id:
+          data.reduce((prev, current) =>
+            prev.id > current.id ? prev : current
+          ).id + 1,
+        name: this.state.name,
+        username: this.state.username,
+        email: this.state.email,
+        phone: this.state.phone
+      }
+    );
     this.setState({
       name: "",
       username: "",
@@ -79,7 +100,8 @@ export default class ApiTest extends React.Component {
 
   handleEdit = e => {
     let { data } = this.state;
-    if (e.target.value === "save") {
+    let pos = data.findIndex(ele => ele.id === Number(e.target.id));
+    if (e.target.value === "save" && window.confirm("confirm updation?")) {
       if (
         this.state.newName === "" ||
         this.state.newUsername === "" ||
@@ -89,7 +111,7 @@ export default class ApiTest extends React.Component {
         alert("field cannot be empty");
         return;
       }
-      let newId = data[e.target.id].id;
+      let newId = data[pos].id;
       /******** api for data updation ****************************/
       fetch("https://jsonplaceholder.typicode.com/users/" + newId, {
         method: "PUT",
@@ -108,7 +130,7 @@ export default class ApiTest extends React.Component {
         .then(data => console.log(data))
         .catch(console.log);
       /*******************************************************/
-      data.splice(e.target.id, 1, {
+      data.splice(pos, 1, {
         id: newId,
         name: this.state.newName,
         username: this.state.newUsername,
@@ -120,29 +142,42 @@ export default class ApiTest extends React.Component {
         data,
         editable: ""
       });
+    } else if (e.target.value === "X") {
+      this.setState({ editable: "" });
     } else {
       this.setState({
-        newName: data[e.target.id].name,
-        newUsername: data[e.target.id].username,
-        newEmail: data[e.target.id].email,
-        newPhone: data[e.target.id].phone,
+        newName: data[pos].name,
+        newUsername: data[pos].username,
+        newEmail: data[pos].email,
+        newPhone: data[pos].phone,
         editable: e.target.id
       });
     }
   };
 
   handleDelete = e => {
+    if (window.confirm("ready to remove?")) {
+      let { data } = this.state;
+      let pos = data.findIndex(ele => ele.id === Number(e.target.id));
+      let newId = data[pos].id;
+      /******** api for data deletion ****************************/
+      fetch("https://jsonplaceholder.typicode.com/users/" + newId, {
+        method: "DELETE"
+      })
+        .then(response => response.json())
+        .then(data => console.log(data))
+        .catch(console.log);
+      /*******************************************************/
+      data.splice(pos, 1);
+      this.setState({
+        data
+      });
+    }
+  };
+
+  handleSort = e => {
     let { data } = this.state;
-    let newId = data[e.target.id].id;
-    /******** api for data deletion ****************************/
-    fetch("https://jsonplaceholder.typicode.com/users/" + newId, {
-      method: "DELETE"
-    })
-      .then(response => response.json())
-      .then(data => console.log(data))
-      .catch(console.log);
-    /*******************************************************/
-    data.splice(e.target.id, 1);
+    data.reverse();
     this.setState({
       data
     });
@@ -167,7 +202,7 @@ export default class ApiTest extends React.Component {
             data={this.state}
           />
         </div>
-        {this.state.data.length > 0 && (
+        {this.state.data.length > 0 ? (
           <div className="UsersTable">
             <Table
               data={data}
@@ -175,8 +210,17 @@ export default class ApiTest extends React.Component {
               handleEdit={this.handleEdit}
               handleChange={this.handleChange}
               editable={this.state.editable}
+              handleSort={this.handleSort}
               newData={this.state}
             />
+          </div>
+        ) : (
+          <div
+            className="emptyList"
+            style={{ width: "80%", margin: "50px auto", textAlign: "center" }}
+          >
+            <Loader />
+            <strong>List is empty. Fill the above form to add users.</strong>
           </div>
         )}
       </div>
